+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * unreleased
   - internal: looser yaml dep for compat with other packages
+  - bugfix: don't drop columns on sqlite (because sqlite doesn't drop columns)
 * 0.1.1
   - bugfix: don't crash in `index_name` when index is unnamed (though anon indexes may cause other problems downstream)
   - bugfix: sa\_harness was broken